[rollout] refactor: rename internal ATIF module
